+++ src/__init__.py
@@ -1,0 +1,1 @@
+# BMI-SOFT EEG Decoding — source package
