+++ src/data/simplify_labels.py
