+++ src/evaluation/__init__.py
@@ -1,0 +1,5 @@
+"""Evaluation and model comparison."""
+
+from src.evaluation.compare_training import main as run_compare_training
+
+__all__ = ["run_compare_training"]
