+++ src/models/__init__.py
@@ -1,0 +1,15 @@
+"""Models and training pipeline: featureizers, classifiers, evaluation."""
+
+from src.models.pipeline import (
+    build_pipeline,
+    load_dataset,
+    load_npz_bundle,
+    evaluate_and_report,
+)
+
+__all__ = [
+    "build_pipeline",
+    "load_dataset",
+    "load_npz_bundle",
+    "evaluate_and_report",
+]
